Create Password Based Door Lock System
The password-based door lock system kit is a device that has been developed to enhance 
the security of a place concerned where the entry of unauthorized individuals is restricted. 
The device works with the help of a microcontroller that has been set in such a way to 
respond to a certain binary code only.